Restructure Project Layout
Takes sub-charts and places them in the correct directory
of the parent chart. Without this change global variables
cannot be shared which leads to re-declaration in every
sub-chart. The same holds true for helper templates.

Also removes any external dependencies that are packed in
the tgz format. If needed they can be downloaded ad-hoc.