Use Scope for Component Name Template
Makes proper use of the scope for the named template
that renders a full name of a component.
This is done to include top-level built-in variables,
objects etc. as well as user-defined ones.
User defined variables, objects etc. can be specified
using a combination of helm's merge and dict functions.