Provision Custom Certificates
Adds a routine that picks up custom certificates from
a pre-defined directory and provisions them in a bundled
form as a ConfigMap.

Only certificates in PEM format are taken into account
(requires .pem file extension). However, other than the
file extension no other check is in place to enforce it.
That leaves it to the admin to provide certificates in
the correct format.

Due to k8s constraints of ConfigMaps the total size of
all certificates must not be larger than 1MB. Also, this
is not enforced by any checks, yet.